Fixed database bugs
